aidoll-pi/funny_chatbot_demo/requirements.txt
boto3
load_dotenv